+++ tencentcloud/antiddos/v20200309/models.go
@@ -7208,7 +7208,7 @@ type L4RuleSource struct {
 	// 回源IP或域名
 	Source *string `json:"Source,omitempty" name:"Source"`
 
-	// 权重值，取值[0,100]
+	// 权重值，取值[0,100]，暂不支持
 	Weight *uint64 `json:"Weight,omitempty" name:"Weight"`
 
 	// 8000