+++ tencentcloud/cbs/v20170312/client.go
@@ -2486,15 +2486,11 @@ func NewModifyDisksChargeTypeResponse() (response *ModifyDisksChargeTypeResponse
 }
 
 // ModifyDisksChargeType
-// 接口请求域名： cbs.tencentcloudapi.com 。
+// 本接口 (ModifyDisksChargeType) 用于切换云硬盘的计费模式。
 //
 // 
 //
-// 本接口 (ModifyDisksChargeType) 用于切换云盘的计费模式。
-//
-// 
-//
-// 非弹性云盘不支持此接口，请通过修改实例计费模式接口将实例连同非弹性云盘一起转换。
+// 非弹性云硬盘不支持此接口，请通过修改实例计费模式接口将实例连同非弹性云硬盘一起转换。
 //
 // 默认接口请求频率限制：10次/秒。
 //
@@ -2517,15 +2513,11 @@ func (c *Client) ModifyDisksChargeType(request *ModifyDisksChargeTypeRequest) (r
 }
 
 // ModifyDisksChargeType
-// 接口请求域名： cbs.tencentcloudapi.com 。
+// 本接口 (ModifyDisksChargeType) 用于切换云硬盘的计费模式。
 //
 // 
 //
-// 本接口 (ModifyDisksChargeType) 用于切换云盘的计费模式。
-//
-// 
-//
-// 非弹性云盘不支持此接口，请通过修改实例计费模式接口将实例连同非弹性云盘一起转换。
+// 非弹性云硬盘不支持此接口，请通过修改实例计费模式接口将实例连同非弹性云硬盘一起转换。
 //
 // 默认接口请求频率限制：10次/秒。
 //