+++ tencentcloud/cbs/v20170312/models.go
@@ -809,6 +809,7 @@ func (r *CreateDisksRequest) FromJsonString(s string) error {
 // Predefined struct for user
 type CreateDisksResponseParams struct {
 	// 创建的云硬盘ID列表。
+	// 注意：此字段可能返回 null，表示取不到有效值。
 	DiskIdSet []*string `json:"DiskIdSet,omitempty" name:"DiskIdSet"`
 
 	// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
@@ -3174,7 +3175,7 @@ func (r *ModifyDiskExtraPerformanceResponse) FromJsonString(s string) error {
 
 // Predefined struct for user
 type ModifyDisksChargeTypeRequestParams struct {
-	// 一个或多个待操作的云硬盘ID。每次请求批量云盘上限为100。
+	// 一个或多个待操作的云硬盘ID。每次请求批量云硬盘上限为100。
 	DiskIds []*string `json:"DiskIds,omitempty" name:"DiskIds"`
 
 	// 设置为预付费模式，即包年包月相关参数设置。通过该参数可以指定包年包月实例的购买时长、是否设置自动续费等属性。
@@ -3187,7 +3188,7 @@ type ModifyDisksChargeTypeRequestParams struct {
 type ModifyDisksChargeTypeRequest struct {
 	*tchttp.BaseRequest
 	
-	// 一个或多个待操作的云硬盘ID。每次请求批量云盘上限为100。
+	// 一个或多个待操作的云硬盘ID。每次请求批量云硬盘上限为100。
 	DiskIds []*string `json:"DiskIds,omitempty" name:"DiskIds"`
 
 	// 设置为预付费模式，即包年包月相关参数设置。通过该参数可以指定包年包月实例的购买时长、是否设置自动续费等属性。
@@ -3378,27 +3379,27 @@ func (r *ModifySnapshotAttributeResponse) FromJsonString(s string) error {
 
 // Predefined struct for user
 type ModifySnapshotsSharePermissionRequestParams struct {
+	// 快照ID, 可通过[DescribeSnapshots](https://cloud.tencent.com/document/api/362/15647)查询获取。
+	SnapshotIds []*string `json:"SnapshotIds,omitempty" name:"SnapshotIds"`
+
 	// 接收分享快照的账号Id列表，array型参数的格式可以参考[API简介](https://cloud.tencent.com/document/api/213/568)。帐号ID不同于QQ号，查询用户帐号ID请查看[帐号信息](https://console.cloud.tencent.com/developer)中的帐号ID栏。
 	AccountIds []*string `json:"AccountIds,omitempty" name:"AccountIds"`
 
 	// 操作，包括 SHARE，CANCEL。其中SHARE代表分享操作，CANCEL代表取消分享操作。
 	Permission *string `json:"Permission,omitempty" name:"Permission"`
-
-	// 快照ID, 可通过[DescribeSnapshots](https://cloud.tencent.com/document/api/362/15647)查询获取。
-	SnapshotIds []*string `json:"SnapshotIds,omitempty" name:"SnapshotIds"`
 }
 
 type ModifySnapshotsSharePermissionRequest struct {
 	*tchttp.BaseRequest
 	
+	// 快照ID, 可通过[DescribeSnapshots](https://cloud.tencent.com/document/api/362/15647)查询获取。
+	SnapshotIds []*string `json:"SnapshotIds,omitempty" name:"SnapshotIds"`
+
 	// 接收分享快照的账号Id列表，array型参数的格式可以参考[API简介](https://cloud.tencent.com/document/api/213/568)。帐号ID不同于QQ号，查询用户帐号ID请查看[帐号信息](https://console.cloud.tencent.com/developer)中的帐号ID栏。
 	AccountIds []*string `json:"AccountIds,omitempty" name:"AccountIds"`
 
 	// 操作，包括 SHARE，CANCEL。其中SHARE代表分享操作，CANCEL代表取消分享操作。
 	Permission *string `json:"Permission,omitempty" name:"Permission"`
-
-	// 快照ID, 可通过[DescribeSnapshots](https://cloud.tencent.com/document/api/362/15647)查询获取。
-	SnapshotIds []*string `json:"SnapshotIds,omitempty" name:"SnapshotIds"`
 }
 
 func (r *ModifySnapshotsSharePermissionRequest) ToJsonString() string {
@@ -3413,9 +3414,9 @@ func (r *ModifySnapshotsSharePermissionRequest) FromJsonString(s string) error {
 	if err := json.Unmarshal([]byte(s), &f); err != nil {
 		return err
 	}
+	delete(f, "SnapshotIds")
 	delete(f, "AccountIds")
 	delete(f, "Permission")
-	delete(f, "SnapshotIds")
 	if len(f) > 0 {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "ModifySnapshotsSharePermissionRequest has unknown keys!", "")
 	}
@@ -3567,7 +3568,7 @@ type Price struct {
 
 // Predefined struct for user
 type RenewDiskRequestParams struct {
-	// 预付费模式，即包年包月相关参数设置。通过该参数可以指定包年包月云盘的续费时长。<br>在云盘与挂载的实例一起续费的场景下，可以指定参数CurInstanceDeadline，此时云盘会按对齐到实例续费后的到期时间来续费。
+	// 预付费模式，即包年包月相关参数设置。通过该参数可以指定包年包月云硬盘的续费时长。<br>在云硬盘与挂载的实例一起续费的场景下，可以指定参数CurInstanceDeadline，此时云硬盘会按对齐到实例续费后的到期时间来续费。
 	DiskChargePrepaid *DiskChargePrepaid `json:"DiskChargePrepaid,omitempty" name:"DiskChargePrepaid"`
 
 	// 云硬盘ID， 通过[DescribeDisks](/document/product/362/16315)接口查询。
@@ -3577,7 +3578,7 @@ type RenewDiskRequestParams struct {
 type RenewDiskRequest struct {
 	*tchttp.BaseRequest
 	
-	// 预付费模式，即包年包月相关参数设置。通过该参数可以指定包年包月云盘的续费时长。<br>在云盘与挂载的实例一起续费的场景下，可以指定参数CurInstanceDeadline，此时云盘会按对齐到实例续费后的到期时间来续费。
+	// 预付费模式，即包年包月相关参数设置。通过该参数可以指定包年包月云硬盘的续费时长。<br>在云硬盘与挂载的实例一起续费的场景下，可以指定参数CurInstanceDeadline，此时云硬盘会按对齐到实例续费后的到期时间来续费。
 	DiskChargePrepaid *DiskChargePrepaid `json:"DiskChargePrepaid,omitempty" name:"DiskChargePrepaid"`
 
 	// 云硬盘ID， 通过[DescribeDisks](/document/product/362/16315)接口查询。