+++ tencentcloud/cynosdb/v20190107/errors.go
@@ -188,6 +188,9 @@ const (
 	// 找不到所选子网。
 	INVALIDPARAMETERVALUE_SUBNETNOTFOUND = "InvalidParameterValue.SubnetNotFound"
 
+	// 未找到该数据。
+	INVALIDPARAMETERVALUE_VALUENOTFOUND = "InvalidParameterValue.ValueNotFound"
+
 	// 找不到所选VPC网络。
 	INVALIDPARAMETERVALUE_VPCNOTFOUND = "InvalidParameterValue.VpcNotFound"