+++ tencentcloud/vpc/v20170312/models.go
@@ -23660,6 +23660,7 @@ type Route struct {
 	GatewayType *string `json:"GatewayType,omitempty" name:"GatewayType"`
 
 	// 下一跳地址，这里只需要指定不同下一跳类型的网关ID，系统会自动匹配到下一跳地址。
+	// 特殊说明：GatewayType为NORMAL_CVM时，GatewayId填写实例的内网IP。
 	GatewayId *string `json:"GatewayId,omitempty" name:"GatewayId"`
 
 	// 路由策略ID。IPv4路由策略ID是有意义的值，IPv6路由策略是无意义的值0。后续建议完全使用字符串唯一ID `RouteItemId`操作路由策略。